+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/PluginSystem/PluginInfo.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: e328efaa3b124b15a3438c0046973e3a
+timeCreated: 1637715379
